코틀린 이관작업 3 일차

+++ src/main/java/org/alham/alhamfirst/dto/user/UserDTO.kt
@@ -19,7 +19,7 @@ data class UserDTO(
         var age : Int = 0,
         var userType : UserType = UserType.BASIC,
         var email : String = "",
-        var statData : MutableMap<String,Int> = mutableMapOf()
+        var statData : MutableMap<String,Long> = mutableMapOf()
 ) {
